move get_number_of_forms_by_type and tests to analytics module
