+++ corehq/ex-submodules/couchforms/dbaccessors.py
@@ -59,19 +59,6 @@ def get_forms_of_all_types(domain):
     ).all()
 
 
-def get_number_of_forms_by_type(domain, type_):
-    assert type_ in doc_types()
-    startkey = [domain, type_]
-    endkey = startkey + [{}]
-    submissions = XFormInstance.view(
-        "couchforms/all_submissions_by_domain",
-        startkey=startkey,
-        endkey=endkey,
-        reduce=True,
-    ).one()
-    return submissions['value'] if submissions else 0
-
-
 def get_number_of_forms_all_domains_in_couch():
     """
     Return number of non-error, non-log forms total across all domains
