+++ corehq/ex-submodules/couchforms/tests/test_dbaccessors.py
@@ -2,7 +2,7 @@ import datetime
 from django.test import TestCase
 
 from corehq.form_processor.tests.utils import FormProcessorTestUtils
-from couchforms.dbaccessors import get_forms_by_type, get_number_of_forms_by_type, \
+from couchforms.dbaccessors import get_forms_by_type, \
     get_form_ids_by_type, get_number_of_forms_all_domains_in_couch
 from couchforms.models import XFormInstance, XFormError
 
@@ -46,18 +46,6 @@ class TestDBAccessors(TestCase):
                          {form._id for form in self.xform_errors})
         for form in forms:
             self.assertIsInstance(form, XFormError)
-
-    def test_get_number_of_forms_by_type_xforminstance(self):
-        self.assertEqual(
-            get_number_of_forms_by_type(self.domain, 'XFormInstance'),
-            len(self.xforms)
-        )
-
-    def test_get_number_of_forms_by_type_xformerror(self):
-        self.assertEqual(
-            get_number_of_forms_by_type(self.domain, 'XFormError'),
-            len(self.xform_errors)
-        )
 
     def test_get_form_ids_by_type(self):
         form_ids = get_form_ids_by_type(self.domain, 'XFormError')
